Create event with macro
